Modified search and trend api
