+++ routes/authRoute.js
@@ -1,3 +1,5 @@
+"use strict";
+
 const passport = require('passport');
 
 module.exports = (app) => {
